refactor: split record validation from postgres adapter
